ref: move from core to cli
